Pass custom nets in to DCC and centralize network definition
Instead of the default/pretrained autoencoder, can pass something else.
Choose a central location to define problem specific network
architecture (data_params).
