+++ pytorch/extract_feature.py
@@ -16,8 +16,6 @@ import torch
 from torch.autograd import Variable
 import torch.backends.cudnn as cudnn
 
-from extractSDAE import extract_sdae_mnist, extract_sdae_reuters, extract_sdae_ytf, extract_sdae_coil100, extract_sdae_yale, extract_sdae_easy
-from extractconvSDAE import extract_convsdae_mnist, extract_convsdae_coil100, extract_convsdae_ytf, extract_convsdae_yale
 from custom_data import DCCPT_data
 
 # Parse all the input argument
@@ -31,7 +29,7 @@ parser.add_argument('--deviceID', type=int, help='deviceID', default=0)
 parser.add_argument('--dim', type=int, help='dimension', default=10)
 parser.add_argument('--h5', dest='h5', help='to store as h5py file', default=False, type=bool)
 
-def main(args):
+def main(args, net=None):
     datadir = get_data_dir(args.db)
     outputdir = get_output_dir(args.db)
 
@@ -46,45 +44,28 @@ def main(args):
         torch.backends.cudnn.enabled = True
         cudnn.benchmark = True
 
-    reluslope = 0.0
     kwargs = {'num_workers': 0, 'pin_memory': True} if use_cuda else {}
     trainset = DCCPT_data(root=datadir, train=True, h5=args.h5)
     testset = DCCPT_data(root=datadir, train=False, h5=args.h5)
-    if args.db == 'mnist':
-        net = extract_sdae_mnist(slope=reluslope, dim=args.dim)
-    elif args.db == 'reuters' or args.db == 'reuters10k' or args.db == 'rcv1':
-        net = extract_sdae_reuters(slope=reluslope, dim=args.dim)
-    elif args.db == 'ytf':
-        net = extract_sdae_ytf(slope=reluslope, dim=args.dim)
-    elif args.db == 'coil100':
-        net = extract_sdae_coil100(slope=reluslope, dim=args.dim)
-    elif args.db == 'yale':
-        net = extract_sdae_yale(slope=reluslope, dim=args.dim)
-    elif args.db == 'cmnist':
-        net = extract_convsdae_mnist(slope=reluslope)
-    elif args.db == 'ccoil100':
-        net = extract_convsdae_coil100(slope=reluslope)
-    elif args.db == 'cytf':
-        net = extract_convsdae_ytf(slope=reluslope)
-    elif args.db == 'cyale':
-        net = extract_convsdae_yale(slope=reluslope)
-    elif args.db == dp.easy.name:
-        net = extract_sdae_easy(slope=reluslope, dim=args.dim)
-    else:
-        raise ValueError("Unexpected database %s" % args.db)
+
+    # load from checkpoint if we're not given an external net
+    load_checkpoint = True if net is None else False
+    if net is None:
+        net = dp.load_predefined_extract_net(args)
 
     totalset = torch.utils.data.ConcatDataset([trainset, testset])
     dataloader = torch.utils.data.DataLoader(totalset, batch_size=100, shuffle=False, **kwargs)
 
     # copying model params from checkpoint
-    filename = os.path.join(outputdir, args.torchmodel)
-    if os.path.isfile(filename):
-        print("==> loading params from checkpoint '{}'".format(filename))
-        checkpoint = torch.load(filename)
-        net.load_state_dict(checkpoint['state_dict'])
-    else:
-        print("==> no checkpoint found at '{}'".format(filename))
-        raise
+    if load_checkpoint:
+        filename = os.path.join(outputdir, args.torchmodel)
+        if os.path.isfile(filename):
+            print("==> loading params from checkpoint '{}'".format(filename))
+            checkpoint = torch.load(filename)
+            net.load_state_dict(checkpoint['state_dict'])
+        else:
+            print("==> no checkpoint found at '{}'".format(filename))
+            raise ValueError
 
     if use_cuda:
         net.cuda()
